requirements.txt
streamlit
diffusers[torch]
torch
Pillow
gradio
scipy
numpy
matplotlib
pandas
seaborn
scikit-learn
xgboost
lightgbm
catboost
statsmodels
nltk
spacy
transformers
beautifulsoup4
requests
flask
django
sqlalchemy
pydantic
fastapi
uvicorn
jupyter
notebook
pytest
black
isort
mypy
bandit
flake8
coverage
Sphinx
alembic
celery
redis
boto3
google-cloud-storage
azure-storage-blob
python-dotenv
pillow
imageio
opencv-python
tqdm
pyyaml
joblib
seaborn
plotly
dash
streamlit
diffusers[torch]
torch
Pillow
